requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
google-genai>=1.0.0
edge-tts>=6.1.0
pydub>=0.25.1
feedgen>=1.0.0
python-dotenv>=1.0.0
PyGithub>=2.0.0
